refactor: split rest common module
